fix(temperature)!:wrong return type & fix temperature calculate formula
-change return type from uint8_t to float
-improve calculation formula

+++ temperature/temperature.h
@@ -45,4 +45,4 @@ void TEMPERATURE_Init(void);
 
  @return  The data, read from sensor
 */
-uint8_t TEMPERATURE_Read(void);
+float TEMPERATURE_Read(void);
